Export logs for developer analysis
